% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP6VisR-package.R
\docType{package}
\name{CMIP6VisR-package}
\alias{CMIP6VisR}
\alias{CMIP6VisR-package}
\title{TVisualization and Analysis of CMIP6 Hydroclimatic Data}
\description{
Provides data manipulation, analyses and visualization for CMIP6 hydroclimatic data. The files are been archived in the Federated Research Data Repository (FRDR; https://doi.org/10.20383/103.0829).
}
\references{
Rajulapati, C., Abdelmoaty, H., Nerantzaki, S., Papalexiou, S. (2024). High-resolution future temperature and precipitation dataset for Canada, 2015 - 2100. Federated Research Data Repository. https://doi.org/10.20383/103.0829
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TycheLab/CMIP6VisR}
}

}
\author{
\strong{Maintainer}: Kevin Shook \email{kshook@kshook.ca} [contributor]

Authors:
\itemize{
  \item Simon Michael Papalexiou
  \item Heba Abdelmoaty
  \item Konstantinos Andreadis
}

}
\keyword{package}
