% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_basin_daily_precip.R
\name{cv_basin_daily_precip}
\alias{cv_basin_daily_precip}
\title{Calculates basin-averaged daily precipitation}
\usage{
cv_basin_daily_precip(
  netcdf_directory = ".",
  scenario = "pr_day_ACCESS-CM2_ssp126_r2i1p1f1_gn_20150101-21001231_cannc_SPQM_",
  basin_zone_area = NULL,
  temp_file = TRUE
)
}
\arguments{
\item{netcdf_directory}{Required. Directory containing NetCDF files.}

\item{scenario}{Required. Full name of scenario to be used. This is the file name omitting the zone number.}

\item{basin_zone_area}{Required. A list object returned by \code{cv_clip_basin} which contains the zone numbers to be used,
the basin area within each zone, and rasters of each zone containing the area of each
element.}

\item{temp_file.}{If \code{TRUE} (the default) then temporary files will be used when}
}
\value{
Reruns a data frame with 2 columns: \code{date} and \code{precipitation}.
}
\description{
A short description...
}
\examples{
\donttest{
az_raster <- cv_zone_area_raster()
fpath <- system.file("extdata", "07BF001.shp", package = "CMIP6VisR")
basin_vector <- vect(fpath)
basin_areas <- cv_clip_basin(az_raster, basin_vector)
netcdf_directory <- "."
values <- cv_basin_daily_precip(netcdf_directory = netcdf_directory,
                               basin_zone_area = basin_areas,
                               temp_file = FALSE)
}
}
\seealso{
\code{\link{cv_clip_basin}}
}
\author{
Kevin Shook
}
