% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_extract_all_basin_daily.R
\name{cv_extract_all_basin_daily}
\alias{cv_extract_all_basin_daily}
\title{Extracts daily precipitation or temperatures for all locations in a basin}
\usage{
cv_extract_all_basin_daily(
  netcdf_directory = ".",
  scenario = "pr_day_ACCESS-CM2_ssp126_r2i1p1f1_gn_20150101-21001231_cannc_SPQM_",
  basin_zone_area = NULL,
  temp_file = TRUE,
  output_file_name = NULL
)
}
\arguments{
\item{netcdf_directory}{Required. Directory containing NetCDF files.}

\item{scenario}{Required. Full name of scenario to be used. This is the file
name omitting the zone number. The first character of the scenario name is used to
determine the name of the variable returned.}

\item{basin_zone_area}{Required. A list object returned by \code{cv_clip_basin()}
which contains the zone numbers to be used, the basin area within each zone,
and rasters of each zone containing the area of each
element.}

\item{temp_file}{If \code{TRUE} (the default) then temporary files will be used when extracting the values
from the NetCDF files. This option is slower than keeping all the values in
memory (which is what occurs if \code{temp_file = TRUE}), \emph{but} allows the function
to work with very large basins, which may require more memory than is available.}

\item{output_file_name}{If \code{NULL}, then the precipitation values
will not be written to a file. If a file name is specified, then the values \emph{will}
be written to a file. Note that if the filename exists, then the values will be
overwritten.}
}
\value{
Returns a \code{SpatRaster} object of the daily precipitation for all
locations in the specified basin. Optionally writes the precipitation values to
a specified NetCDF file.
}
\description{
Extracts the daily precipitation or temperature values from CMIP6 NetCDF files for all locations
in a specified basin.

The CMIP6 data are arranged by zone in 9 files. As a given basin may lie over more than
one zone, it may necessary to read in data from more than one NetCDF file when computing
the basin mean precipitation.
}
\examples{
\donttest{
# This function is marked to not be tested as it requires local copies of the CMIP6 data files
# which are many GB in size
library(terra)
az_raster <- cv_zone_area_raster()
fpath <- system.file("extdata", "07BF001.shp", package = "CMIP6VisR")
basin_vector <- vect(fpath)
basin_areas <- cv_clip_basin(az_raster, basin_vector)
netcdf_directory <- "."
all_precip <- cv_extract_all_basin_daily(netcdf_directory = netcdf_directory,
                               basin_zone_area = basin_areas,
                               temp_file = FALSE)
}
}
\seealso{
\code{\link{cv_clip_basin}} \code{\link{cv_extract_mean_basin_daily}}
}
\author{
Kevin Shook
}
