% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_extract_point_daily.R
\name{cv_extract_point_daily}
\alias{cv_extract_point_daily}
\title{Extracts daily precipitation or temperature for a specific location}
\usage{
cv_extract_point_daily(
  variable_file_name = NULL,
  grid_lon = NULL,
  grid_lat = NULL
)
}
\arguments{
\item{variable_file_name}{Required. String containing the name of the NetCDF
file with the CMIP6 daily precipitation or temperature, including the file path.}

\item{grid_lon}{Required. Longitude of the grid cell containing the location,
as returned by \code{cv_find_point}.}

\item{grid_lat}{Required. Latitude of grid cell containing the location,
as returned by \code{cv_find_point}.}
}
\value{
Returns a data frame with 2 columns: \code{date} and either \code{precipitation}
or \code{temperature}, depending on the file name. The
\code{date} is a standard \R date over the interval 2015-01-01 to 2100-12-31, and the
\code{precipitation} or \code{temperature} is the value at the specified grid location. The returned time series
can be plotted using
\code{cv_plot_TS}, \code{cv_plot_prob}, and \code{cv_plot_season}.
}
\description{
Extracts a daily time series from a single CMIP6 raster location, which can
be used to compare the CMIP6 values against station values. The location
of the data in the CMIP6 data (its zone and raster location) is found using
\code{cv_find_point}.
}
\examples{
\dontrun{
point_long <- -110.456
point_lat <- 52.19
point_loc <- cv_find_point(point_long, point_lat)
point_loc
precip_file <- "pr_day_ACCESS-CM2_ssp126_r2i1p1f1_gn_20150101-21001231_cannc_SPQM_07.nc"
grid_lon <- point_loc[1]
grid_lat <- point_loc[2]
point_precip <- cv_extract_point_daily(precip_file, grid_lon, grid_lat)
}

}
\seealso{
\code{\link{cv_find_point}} \code{\link{cv_basin_daily_precip}}
}
\author{
Kevin Shook
}
