% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_map.R
\name{cv_plot_map}
\alias{cv_plot_map}
\title{Plot Spatial Precipitation or Temperature Distribution}
\usage{
cv_plot_map(nc_file, stat_type = "daily", variable = "precipitation")
}
\arguments{
\item{nc_file}{A path to a NetCDF (.nc) file readable by the \code{stars} package.}

\item{stat_type}{One of "daily", "monthly", or "annual" to indicate which type of statistic to visualize.
"daily" provides the nonzero average of daily values during the whole period given in the NC file.
"monthly" provides the average monthly totals (precipitation) or means (temperature).
"annual" provides the annual maxima of daily values during the whole duration.}

\item{variable}{A string starting with 'p' (precipitation) or 't' (temperature).}
}
\value{
A ggplot object showing the specified variable's distribution over space.
Note: We are using the default Viridis color scale which can be overwritten.
The returned plots look best when saved at the size 14 x 12 cm.
You can easily change the font sizes using theme().
}
\description{
This function takes a NetCDF file containing daily time series of precipitation or temperature,
extracts spatial and temporal features, and plots one of three options: daily nonzero frequency,
average monthly totals, or average annual maxima.
}
\examples{
cv_plot_map("Pincher_ck.nc", stat_type = "monthly", variable = "temperature")
cv_plot_map("Pincher_ck.nc", stat_type = "annual", variable = "precip")
}
