% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_prob.R
\name{cv_prob}
\alias{cv_prob}
\title{Plot Probability of Exceedance for Precipitation}
\usage{
cv_prob(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the probability of exceedance of nonzero precipitation.
}
\description{
This function takes a data frame containing a precipitation time series,
filters out zero precipitation values, calculates the probability of exceedance,
and generates a scatter plot with a logarithmic y-axis.
}
\examples{
\dontrun{
data <- readRDS("data/eg_TS.rds")
colnames(data) <- c("date", "precipitation")
cv_prob(data)
#recommended save size
ggsave("plot_prob.jpg", height = 12, width = 16.5, units = "cm", dpi = 400)
}
}
