% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_SeasMon.R
\name{cv_SeasMon}
\alias{cv_SeasMon}
\title{Plot Seasonal Precipitation Distribution}
\usage{
cv_SeasMon(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the seasonal distribution of nonzero precipitation,
with mean values and P0 labeled.
}
\description{
This function takes a data frame with a precipitation time series, assigns seasons based on months,
computes the proportion of zero precipitation values (P0) and seasonal means for nonzero precipitation,
and generates a violin plot showing the seasonal distribution of precipitation.
}
\examples{
\dontrun{
data <- readRDS("data/eg_TS.rds")
colnames(data) <- c("date", "precipitation")
cv_SeasMon(data)
#recommended save size
gggsave("plot_seasALL.jpg", height = 14, width = 16.5, units = "cm", dpi = 400)
}
}
