% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_TS.R
\name{cv_TS}
\alias{cv_TS}
\title{Plot Precipitation Time Series}
\usage{
cv_TS(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the precipitation time series with summary statistics in the title.
}
\description{
This function takes a data frame containing a precipitation time series,
calculates basic statistics (mean, standard deviation, and proportion of zero values),
and generates a ggplot visualization.
}
\examples{
\dontrun{
data <- readRDS("data/eg_TS.rds")
colnames(data) <- c("date", "precipitation")
cv_TS(data)
#Recommended save size
ggsave("plot_TS.jpg", height = 6, width = 16.5, units = "cm", dpi = 400)
}
}
