% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_grid_df.R
\docType{data}
\name{zone_grid_df}
\alias{zone_grid_df}
\title{zone_grid_df}
\format{
A dateframe with 169584 rows and 4 columns.
}
\source{
https://doi.org/10.20383/103.0829
}
\usage{
zone_grid_df
}
\description{
A dataframe of zone values and areas for all CMIP6 grid cells
}
\details{
Variables:
\describe{
\item{long}{Longitude}
\item{lat}{Latitude}
\item{zone}{Zone (0-9)}
\item{area}{Grid cell area km\eqn{^2}{^2}}
}
}
\keyword{datasets}
