% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_all_basin_daily_precip.R,
%   R/cv_concat_daily_precip.R
\name{cv_all_basin_daily_precip}
\alias{cv_all_basin_daily_precip}
\title{Calculate daily mean precipitation for each of 11 Canadian major river basins}
\usage{
cv_all_basin_daily_precip(
  nc_dir = ".",
  filespec = "*.nc",
  out_file = "daily.nc"
)

cv_all_basin_daily_precip(
  nc_dir = ".",
  filespec = "*.nc",
  out_file = "daily.nc"
)
}
\arguments{
\item{nc_dir}{Directory containing daily precipitation NetCDF files}

\item{filespec}{Name of all required files, including wild cards}

\item{out_file}{Output file including path}
}
\value{
Writes all daily mean precipitation values to the specified file.

Writes all daily mean precipitation values to the specified file.
}
\description{
Calculate daily mean precipitation for each of 11 Canadian major river basins

Concatenate daily precipitation files
}
