% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_clip_basin.R
\name{cv_clip_basin}
\alias{cv_clip_basin}
\title{Get zones and areas that clip rasters from a given basin polygon}
\usage{
cv_clip_basin(za_rast, basin)
}
\arguments{
\item{za_rast}{\code{SpatRaster} of the zones and areas}

\item{basin}{\code{SpatVector} of the basin to clip to}
}
\description{
Takes the \code{SpatRaster} that represents zones and areas and
clips them from polygon \code{basin}
}
\author{
Kostas Andreadis
}
