% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_find_point.R
\name{cv_find_point}
\alias{cv_find_point}
\title{Locates a specific point in the CMIP6 rasters}
\usage{
cv_find_point(longitude, latitude, fast = TRUE)
}
\arguments{
\item{longitude}{Required. Longitude (decimal degrees) of the point}

\item{latitude}{Required. Latitude (decimal degrees) of the point}

\item{fast}{Optional. If \code{TRUE} (the default), the faster, but slightly less accurate
\code{distVincentySphere} method will be used to determine the distance between the
coordinates of the specified point and that of all the CMIP6 raster points. If \code{FALSE},
then the more accurate \code{distVincentyEllipsoid} method is used. The difference
between the distances of the two methods is typically very small.}
}
\value{
Returns a vector of raster cell coordinates, zone number, cell area (km) and the distance
between point and cell centre (m).
}
\description{
Finds the closest CMIP6 raster location to a point specified by its longitude
and latitude.
}
