% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_prob.R
\name{cv_plot_prob}
\alias{cv_plot_prob}
\title{Plot Probability of Exceedance for Precipitation}
\usage{
cv_plot_prob(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the probability of exceedance of nonzero precipitation.
The returned plots look best when saved at the size 16.5 x 12 cm.
You can easily change the font sizes using theme().
}
\description{
This function takes a data frame containing a precipitation time series,
filters out zero precipitation values, calculates the probability of exceedance,
and generates a scatter plot with a logarithmic y-axis.
}
\examples{
cv_plot_prob(eg_TS)
}
