% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_prob.R
\name{cv_plot_prob}
\alias{cv_plot_prob}
\title{Plot Probability of Exceedance for Precipitation}
\usage{
cv_plot_prob(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the probability of exceedance of nonzero precipitation.
Note: The recommended aspect ratio and size for saving the plots,
based on the dimensions specified in the plot function, are a width of 16.5 cm and a height of 12 cm.
}
\description{
This function takes a data frame containing a precipitation time series,
filters out zero precipitation values, calculates the probability of exceedance,
and generates a scatter plot with a logarithmic y-axis.
}
\examples{
# Load example data from the package
data("eg_TS")
cv_plot_prob(eg_TS)
}
