% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_season.R
\name{cv_plot_season}
\alias{cv_plot_season}
\title{Plot Seasonal Precipitation Distribution}
\usage{
cv_plot_season(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the seasonal distribution of nonzero precipitation,
with mean values and P0 labeled. Note: The recommended aspect ratio and size for saving the plots,
based on the dimensions specified in the plot function, are a width of 16.5 cm and a height of 14 cm.
}
\description{
This function takes a data frame with a precipitation time series, assigns seasons based on months,
computes the proportion of zero precipitation values (P0) and seasonal means for nonzero precipitation,
and generates a violin plot showing the seasonal distribution of precipitation.
}
\examples{
data("eg_TS")
cv_plot_season(eg_TS)
}
