% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_TS.R
\name{cv_plot_TS}
\alias{cv_plot_TS}
\title{Plot precipitation time series}
\usage{
cv_plot_TS(data)
}
\arguments{
\item{data}{A data frame with two columns: \code{date} (date or datetime) and
\code{precipitation} (numeric) as returned by \code{cv_basin_daily_precip()}.}
}
\value{
A ggplot object displaying the precipitation time series with summary
statistics in the title. The returned plots look best when saved at the size
16.5 x 6 cm You can easily change the font sizes using theme().
}
\description{
Creates a time series plot from a data frame of precipitation data
as returned by \code{cv_basin_daily_precip()}. Basic statistics (mean, standard deviation, and proportion of zero values),
are added to the title of the plot.
}
\examples{
cv_plot_TS(eg_TS)
}
\seealso{
\code{\link{cv_basin_daily_precip}} \code{\link{cv_plot_prob}} \code{\link{cv_plot_season}}
}
