% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_TS.R
\name{cv_plot_TS}
\alias{cv_plot_TS}
\title{Plot Precipitation Time Series}
\usage{
cv_plot_TS(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the precipitation time series with summary statistics in the title.
Note: The recommended aspect ratio and size for saving the plots,
based on the dimensions specified in the plot function, are a width of 16.5 cm and a height of 6 cm.
}
\description{
This function takes a data frame containing a precipitation time series,
calculates basic statistics (mean, standard deviation, and proportion of zero values),
and generates a ggplot visualization.
}
\examples{
data("eg_TS")
cv_plot_TS(eg_TS)
}
