% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_TS.R
\name{cv_plot_TS}
\alias{cv_plot_TS}
\title{Plot Precipitation Time Series}
\usage{
cv_plot_TS(data)
}
\arguments{
\item{data}{A data frame with two columns: "date" (date or datetime) and
"precipitation" (numeric).}
}
\value{
A ggplot object displaying the precipitation time series with summary statistics in the title.
The returned plots look best when saved at the size 16.5 x 6 cm.
You can easily change the font sizes using theme().
}
\description{
This function takes a data frame containing a precipitation time series,
calculates basic statistics (mean, standard deviation, and proportion of zero values),
and generates a ggplot visualization.
}
\examples{
cv_plot_TS(eg_TS)
}
