% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_precip_map.R
\name{plot_precip_map}
\alias{plot_precip_map}
\title{Plot Spatial Precipitation Distribution}
\usage{
plot_precip_map(nc_file, stat_type = "daily")
}
\arguments{
\item{nc_file}{A path to a NetCDF (.nc) file readable by the \code{stars} package.}

\item{stat_type}{One of "daily", "monthly", or "annual" to indicate which type of statistic to visualize.
"daily" provides the nonzero average of daily precipitation during the whole period given in the NC file.
"monthly" provides the average total precipitation within every month during the whole duration.
"annual" provides the annual maxima of daily precipitation during the whole duration.}
}
\value{
A ggplot object showing the specified precipitation distribution over space.Note: We are using the default "Virdis"
color scale which could be overwritten.
The returned plots look best when saved at the size 14 x 12 cm.
You can easily change the font sizes using theme().
}
\description{
This function takes a NetCDF file containing daily precipitation time series, extracts spatial and temporal features,
and plots one of three options: daily nonzero frequency, average monthly totals, or average annual maxima.
}
\examples{
plot_precip_map("Pincher_ck.nc", stat_type = "monthly")
}
