% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_temp_map.R
\name{plot_temp_map}
\alias{plot_temp_map}
\title{Plot Spatial Temperature Distribution}
\usage{
plot_temp_map(nc_file, stat_type = c("daily", "monthly", "annual"))
}
\arguments{
\item{nc_file}{A path to a NetCDF (.nc) file readable by the \code{stars} package.}

\item{stat_type}{One of "daily", "monthly", or "annual" to indicate which type of statistic to visualize.
"daily" provides the nonzero average of daily temperature during the whole period given in the NC file.
"monthly" provides the average Temperature within every month during the whole duration.
"annual" provides the annual maxima of daily average Temperature during the whole duration
(not commonly used in applications).}
}
\value{
A ggplot object showing the specified Temperature distribution over space.Note: We are using the default "plasma"
color scale which could be overwritten.
The returned plots look best when saved at the size 14 x 12 cm.
You can easily change the font sizes using theme().
}
\description{
This function takes a NetCDF file containing daily Temperature time series, extracts spatial and temporal features,
and plots one of three options: daily nonzero frequency, average monthly totals, or average annual maxima.
}
\examples{
plot_precip_map("Pincher_ck.nc", stat_type = "monthly")
}
